✨ 80 warnings eliminados (0 warnings - código limpio)
- rydit-lexer: cargo fix unused variable
- migui: allow dead_code (backend SDL2, fonts)
- rydit-gfx: remove unnecessary unsafe, allow dead_code
- rydit-stream: allow dead_code (lan, portal)
- rydit-vm: cargo fix unused variables
- rydit-rs bins: fix unused imports, variables, assignments
- main.rs: clone()→to_string(), allow static_mut_refs
- rybot: allow dead_code module-wide (registry en desarrollo)
- modules: allow dead_code (features futuras)

Total: 27 archivos, 102 inserciones, 76 eliminaciones

+++ crates/rydit-gfx/src/backend_sdl2.rs
@@ -164,8 +164,8 @@ impl Sdl2Backend {
                     win_event: sdl2::event::WindowEvent::Resized(w, h),
                     ..
                 } => {
-                    self.width = w as i32;
-                    self.height = h as i32;
+                    self.width = w;
+                    self.height = h;
                     let viewport = Rect::new(0, 0, w as u32, h as u32);
                     self.canvas.set_viewport(viewport);
                 }
